Review structs implementing interfaces{} * Refactor payments sources to implement PaymentSource interface * Refactor tokens data to implement TokenData interface

+++ test/payments_request_previous_test.go
@@ -318,7 +318,7 @@ func TestMakePaymentsIdempotentlyPrevious(t *testing.T) {
 
 }
 
-func makeCardPaymentPrevious(t *testing.T, shouldCapture bool, amount int) *abc.PaymentResponse {
+func makeCardPaymentPrevious(t *testing.T, shouldCapture bool, amount int64) *abc.PaymentResponse {
 
 	cardSource := sources.NewRequestCardSource()
 	cardSource.Name = Name
